Complete backup folder layout refactor.
