+++ backup_odoo/dev/Vagrantfile
@@ -24,6 +24,7 @@ Vagrant.configure("2") do |config|
     snapshot.vm.box_download_insecure = true
 
     snapshot.vm.provision :shell, path: "setup_rsnapshot.sh"
+    snapshot.vm.synced_folder "D:/SharedFolder/", "/home/vagrant/SharedFolder"
     #snapshot.vm.provision "shell", run: "always", inline: "cp shared/rsnapshot.conf /etc/"
   end
