+++ backup_odoo/dev/setup_odoo.sh
