+++ backup_odoo/dev/setup_rsnapshot.sh
