+++ backup_odoo/example/conf.json
@@ -1,6 +1,6 @@
 {
     "backup_root": "/home/vagrant/backup/",
-    "conf_path": "/home/vagrant/shared/test.conf",
+    "conf_path": "/home/vagrant/test.conf",
     "databases": [
         {
             "base_url": "vagrant@192.168.50.4",
